docker/drive-api-conn/requirements.txt
Flask
flask-session
google_api_python_client==2.66.0
protobuf==4.25.1
Requests==2.31.0
Werkzeug==2.2.3
redis==5.0.1
